+++ lib/marvel/cli/characters.ex
@@ -3,43 +3,43 @@ defmodule Marvel.CLI.Characters do
   
   alias Marvel.CLI.Utils
 
-  def characters() do
+  def list() do
     Marvel.API.Characters.list()
     |> Utils.process_results(&Utils.format_character_output/1)
   end
 
-  def character_by_name(name) do
+  def by_name(name) do
     Marvel.API.Characters.list(%{name: name})
     |> Utils.process_results(&Utils.format_character_output/1)
   end
 
-  def character_by_id(id) do
+  def by_id(id) do
     Marvel.API.Characters.get(id)
     |> Utils.process_results(&Utils.format_character_output/1)
   end
 
-  def character_search(searchText) do
+  def search(searchText) do
     Marvel.API.Characters.list(%{nameStartsWith: searchText})
     |> Utils.process_results(&Utils.format_character_output/1)
   end
 
-  def character_comics(id) do
+  def comics(id) do
     Marvel.API.Characters.comics(id)
-    |> Utils.process_results(&Utils.format_comic_output/1)
+    |> Utils.process_results(&Utils.format_comic_list_output/1)
   end
 
-  def character_events(id) do
+  def events(id) do
     Marvel.API.Characters.events(id)
-    |> Utils.process_results(&Utils.format_event_output/1)
+    |> Utils.process_results(&Utils.format_event_list_output/1)
   end
 
-  def character_series(id) do
+  def series(id) do
     Marvel.API.Characters.series(id)
-    |> Utils.process_results(&Utils.format_series_output/1)
+    |> Utils.process_results(&Utils.format_series_list_output/1)
   end
 
-  def character_stories(id) do
+  def stories(id) do
     Marvel.API.Characters.stories(id)
-    |> Utils.process_results(&Utils.format_story_output/1)
+    |> Utils.process_results(&Utils.format_story_list_output/1)
   end
 end