+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v1.0.0
+
+* Enhancements
+  * Refactored the Marvel API into modules for each entity
+  * Implemented CLI commands for all the entities
+
 # v0.4.0
 
 * Enhancements